[FEATURE] Add automatic mode
Automatic mode can be enabled in the extension configuration. It is off by default.
If automatic mode is enabled the frontend will behave according to what the page module displays for each language.
When it says "Connected Mode", then "fallback" will be used in the frontend.
When it says "Free Mode", then "free" will be used in the frontend.

With automatic mode enabled, you can still override the behaviour in the translated page properties.